[Onboard]
Query to poast a new sensor to the database. Also associatates
a mac_address with a piece of equipment when doing the post call
. If the mac_address exists, then a new piece of equipment is not
inserted (because one exists), otherwise the new equipment is added
to t_equipment. Next the equipment id is obtained and the new sensor
is inserted into the database